+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 
 ## [1.0.2] 2025-11-24
 - Enhance role fetching to display account names while configuring a profile.
+- Added support for Prism EU instance